# Add a benchmark for the numeric facet builder and use sort.Sort in it (just like for the terms one) 