@mschoch 2015-01-16
Sorry I've been slow to get around to this one.  I think this is the right approach, though I'm not sure the tests do the right thing.

Do you remember why you manually created the nu.PrefixCoded values?  Or what the significance was of the 4 values you chose here?  https://github.com/Shugyousha/bleve/commit/222770b1cce31d61d89008efcef343d470fd9ad9#diff-6f28ca93dc2a04f0ba1465c0b1902f0eR14